requirements.txt
Django==2.0.4
python-decouple==3.1
requests==2.18.4

# tests
pytest==3.5.1
vcrpy==1.11.1
